examples/kdotp_weyl/requirements.txt
phasemap
z2pack
fsc.async-tools
